fix: update db schema

+++ packages/backend/db/migrations/20241230061647-createTablePets/migration.sql
@@ -1,6 +1,6 @@
 -- CreateTable
 CREATE TABLE `pets` (
-    `id` BIGINT NOT NULL AUTO_INCREMENT,
+    `id` INT NOT NULL AUTO_INCREMENT,
 
     PRIMARY KEY (`id`)
 ) DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci;
